Ports storage datumization and expansion of grab interaction. (#910)
* 15120

* 15187 part 1

* 15187 part 2

* 15187

* 15061 part 1

* 15061 final

* 15331

* 15368

* 15739

* 15780

* 15802

* 15826

* 15950

* 15866

* 15985

* 16164

* 16938

* 17174

* Ports CM slam sound effects (#15761)


* Update holsters.dm

* 15950 part 2

* proper naming

* Update autoLabelConfig.js

* hm

* hmm

* Update TGS_Talos.dmm

* bruh

* fix

* Update storage.dm

* Update disposal.dm

* proper dump contents

* that'll do

* Update lockbox.dm

---------

+++ code/modules/clothing/glasses/glasses.dm
@@ -93,27 +93,24 @@
 
 /obj/item/clothing/glasses/eyepatch/attackby(obj/item/I, mob/user, params)
 	. = ..()
-
+	if(.)
+		return
+	var/obj/item/clothing/glasses/P
 	if(istype(I, /obj/item/clothing/glasses/hud/health))
-		var/obj/item/clothing/glasses/hud/medpatch/P = new
+		P = new
 		to_chat(user, span_notice("You fasten the medical hud projector to the inside of the eyepatch."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
 	else if(istype(I, /obj/item/clothing/glasses/meson))
-		var/obj/item/clothing/glasses/meson/eyepatch/P = new
+		P = new
 		to_chat(user, span_notice("You fasten the meson projector to the inside of the eyepatch."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
 	if(istype(I, /obj/item/clothing/glasses/night/imager_goggles))
-		var/obj/item/clothing/glasses/night/imager_goggles/eyepatch/P = new
+		P = new
 		to_chat(user, span_notice("You fasten the optical scanner to the inside of the eyepatch."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
-
-		update_icon()
+	if(!P)
+		return
+	qdel(I)
+	qdel(src)
+	user.put_in_hands(P)
+	update_icon()
 
 /obj/item/clothing/glasses/monocle
 	name = "monocle"
@@ -130,6 +127,8 @@
 
 /obj/item/clothing/glasses/regular/attackby(obj/item/I, mob/user, params)
 	. = ..()
+	if(.)
+		return
 
 	if(istype(I, /obj/item/clothing/glasses/hud/health))
 		var/obj/item/clothing/glasses/hud/medglasses/P = new
@@ -164,20 +163,19 @@
 
 /obj/item/clothing/glasses/mgoggles/attackby(obj/item/our_item, mob/user, params)
 	. = ..()
+	if(.)
+		return
 	if(istype(our_item, /obj/item/clothing/glasses/night/imager_goggles))
+		var/obj/item/clothing/glasses/night/optgoggles/our_glasses
 		if(prescription)
-			var/obj/item/clothing/glasses/night/optgoggles/prescription/our_glasses = new
+			our_glasses = new
 			to_chat(user, span_notice("You fasten the optical imaging scanner to the inside of the goggles."))
-			qdel(our_item)
-			qdel(src)
-			user.put_in_hands(our_glasses)
 		else
-			var/obj/item/clothing/glasses/night/optgoggles/our_glasses = new
+			our_glasses = new
 			to_chat(user, span_notice("You fasten the optical imaging scanner to the inside of the goggles."))
-			qdel(our_item)
-			qdel(src)
-			user.put_in_hands(our_glasses)
-
+		qdel(our_item)
+		qdel(src)
+		user.put_in_hands(our_glasses)
 		update_icon(user)
 
 /obj/item/clothing/glasses/mgoggles/prescription
@@ -187,35 +185,30 @@
 
 /obj/item/clothing/glasses/mgoggles/attackby(obj/item/I, mob/user, params)
 	. = ..()
+	if(.)
+		return
 
+	var/obj/item/clothing/glasses/P
 	if(istype(I, /obj/item/clothing/glasses/hud/health))
 		if(prescription)
-			var/obj/item/clothing/glasses/hud/medgoggles/prescription/P = new
+			P = new
 			to_chat(user, span_notice("You fasten the medical hud projector to the inside of the goggles."))
-			qdel(I)
-			qdel(src)
-			user.put_in_hands(P)
 		else
-			var/obj/item/clothing/glasses/hud/medgoggles/S = new
+			P = new
 			to_chat(user, span_notice("You fasten the medical hud projector to the inside of the goggles."))
-			qdel(I)
-			qdel(src)
-			user.put_in_hands(S)
 	else if(istype(I, /obj/item/clothing/glasses/meson))
 		if(prescription)
-			var/obj/item/clothing/glasses/meson/enggoggles/prescription/P = new
+			P = new
 			to_chat(user, span_notice("You fasten the optical meson scanner to the inside of the goggles."))
-			qdel(I)
-			qdel(src)
-			user.put_in_hands(P)
 		else
-			var/obj/item/clothing/glasses/meson/enggoggles/S = new
+			P = new
 			to_chat(user, span_notice("You fasten the optical meson scanner to the inside of the goggles."))
-			qdel(I)
-			qdel(src)
-			user.put_in_hands(S)
-
-		update_icon()
+	if(!P)
+		return
+	qdel(I)
+	qdel(src)
+	user.put_in_hands(P)
+	update_icon()
 
 /obj/item/clothing/glasses/m42_goggles
 	name = "\improper M42 scout sight"
@@ -337,33 +330,28 @@
 
 /obj/item/clothing/glasses/sunglasses/fake/attackby(obj/item/I, mob/user, params)
 	. = ..()
+	if(.)
+		return
 
+	var/obj/item/clothing/glasses/P
 	if(istype(I, /obj/item/clothing/glasses/hud/health))
-		var/obj/item/clothing/glasses/hud/medsunglasses/P = new
+		P = new
 		to_chat(user, span_notice("You fasten the medical hud projector to the inside of the glasses."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
 	else if(istype(I, /obj/item/clothing/glasses/meson))
-		var/obj/item/clothing/glasses/meson/sunglasses/P = new
+		P = new
 		to_chat(user, span_notice("You fasten the optical meson scaner to the inside of the glasses."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
 	else if(istype(I, /obj/item/clothing/glasses/night/m56_goggles))
-		var/obj/item/clothing/glasses/night/sunglasses/P = new
+		P = new
 		to_chat(user, span_notice("You fasten the KTLD sight to the inside of the glasses."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
 	else if(istype(I, /obj/item/clothing/glasses/night/imager_goggles))
-		var/obj/item/clothing/glasses/night/imager_goggles/sunglasses/P = new
+		P = new
 		to_chat(user, span_notice("You fasten the optical imager scaner to the inside of the glasses."))
-		qdel(I)
-		qdel(src)
-		user.put_in_hands(P)
-
-		update_icon()
+	if(!P)
+		return
+	qdel(I)
+	qdel(src)
+	user.put_in_hands(P)
+	update_icon()
 
 /obj/item/clothing/glasses/sunglasses/fake/prescription
 	name = "prescription sunglasses"
@@ -435,27 +423,24 @@
 
 /obj/item/clothing/glasses/orange/attackby(obj/item/our_item, mob/user, params)
 	. = ..()
+	if(.)
+		return
+	var/obj/item/clothing/glasses/our_glasses
 	if(istype(our_item, /obj/item/clothing/glasses/hud/health))
-		var/obj/item/clothing/glasses/hud/orange_glasses/our_glasses = new
+		our_glasses = new
 		to_chat(user, span_notice("You fasten the medical hud projector to the inside of the glasses."))
-		qdel(our_item)
-		qdel(src)
-		user.put_in_hands(our_glasses)
-		update_icon()
 	else if(istype(our_item, /obj/item/clothing/glasses/night/imager_goggles))
-		var/obj/item/clothing/glasses/night/imager_goggles/orange_glasses/our_glasses = new
+		our_glasses = new
 		to_chat(user, span_notice("You fasten the optical imager scaner to the inside of the glasses."))
-		qdel(our_item)
-		qdel(src)
-		user.put_in_hands(our_glasses)
-		update_icon()
 	else if(istype(our_item, /obj/item/clothing/glasses/meson))
-		var/obj/item/clothing/glasses/meson/orange_glasses/our_glasses = new
+		our_glasses = new
 		to_chat(user, span_notice("You fasten the optical meson scaner to the inside of the glasses."))
-		qdel(our_item)
-		qdel(src)
-		user.put_in_hands(our_glasses)
-		update_icon()
+	if(!our_glasses)
+		return
+	qdel(our_item)
+	qdel(src)
+	user.put_in_hands(our_glasses)
+	update_icon()
 
 /obj/item/clothing/glasses/meson/orange_glasses
 	name = "Orange glasses"
